Kernel: Extract DeferredCallEntry from Arch/Processor.h
